test(refactor): move mock modal to utilities
