+++ src/components/global-nav/mobile/global-nav-mobile-menu.spec.js
@@ -1,4 +1,4 @@
-import { render, fireEvent } from 'test-utils'
+import { render, fireEvent, mockModal } from 'test-utils'
 import '@testing-library/jest-dom/extend-expect'
 import GlobalNavMobileMenu from './global-nav-mobile-menu'
 
@@ -23,20 +23,9 @@ const baseProps = {
 
 const { links } = baseProps
 
-// Mock modal
-const setAppElementStub = jest.fn()
-const ReactModalMock = ({ children }) => <div>{children}</div>
-ReactModalMock.setAppElement = setAppElementStub
-jest.mock('react-modal', () => ReactModalMock)
-
-let portalRoot = document.getElementById('portal')
-if (!portalRoot) {
-  const portalRoot = document.createElement('div')
-  portalRoot.setAttribute('id', 'root')
-  document.body.appendChild(portalRoot)
-}
-
 describe('GlobalNavMobileMenu', () => {
+  mockModal()
+
   it('shows the menu when `isOpen` is true', () => {
     const { getByCy } = render(<GlobalNavMobileMenu {...baseProps} isOpen={true} />)
     expect(getByCy('mobile-menu'))
